[reggen] Add pseudo-markdown support to gen_cfg_html.py
This adds formatting and links to the peripheral pins, interrupts and
alerts tables.
